requirements.txt
openai>=1.0
requests
flask
python-dotenv
twilio
pydub
ffmpeg-python
